Saim Bilal - ML Week 3 Task - STU-ML-251-490/requirements.txt
tensorflow
streamlit
opencv-python
pillow
matplotlib
numpy